// src/app/types/spotify.ts
export interface SpotifyPlaylist {
    id: string
    name: string
    description: string
    images: Array<{
      url: string
      height: number
      width: number
    }>
    tracks: {
      total: number
    }
  }
  
  export interface SpotifyTrack {
    id: string
    name: string
    artists: Array<{
      name: string
    }>
    album: {
      name: string
      images: Array<{
        url: string
        height: number
        width: number
      }>
    }
    duration_ms: number
    uri: string
  }
  
  export interface PlaylistTracksResponse {
    items: Array<{
      track: SpotifyTrack
    }>
  }
  
  declare global {
    interface Window {
      onSpotifyWebPlaybackSDKReady: () => void
      Spotify: {
        Player: new (options: {
          name: string
          getOAuthToken: (cb: (token: string) => void) => void
          volume: number
        }) => SpotifyPlayer
      }
    }
  
    interface SpotifyPlayer {
      addListener: (event: string, callback: (...args: any[]) => void) => boolean
      removeListener: (event: string, callback?: (...args: any[]) => void) => boolean
      connect: () => Promise<boolean>
      disconnect: () => void
      getCurrentState: () => Promise<SpotifyPlaybackState | null>
      getVolume: () => Promise<number>
      nextTrack: () => Promise<void>
      pause: () => Promise<void>
      previousTrack: () => Promise<void>
      resume: () => Promise<void>
      seek: (positionMs: number) => Promise<void>
      setName: (name: string) => Promise<void>
      setVolume: (volume: number) => Promise<void>
      togglePlay: () => Promise<void>
    }
  
    interface SpotifyPlaybackState {
      context: {
        uri: string
        metadata: any
      }
      disallows: {
        pausing: boolean
        peeking_next: boolean
        peeking_prev: boolean
        resuming: boolean
        seeking: boolean
        skipping_next: boolean
        skipping_prev: boolean
      }
      paused: boolean
      position: number
      repeat_mode: number
      shuffle: boolean
      track_window: {
        current_track: SpotifyTrack
        next_tracks: SpotifyTrack[]
        previous_tracks: SpotifyTrack[]
      }
    }
  }